+++ app/resources/numismatics/issues/numismatics/issue_wayfinder.rb
@@ -5,7 +5,7 @@ module Numismatics
     relationship_by_property :members, property: :member_ids
     relationship_by_property :numismatic_places, property: :numismatic_place_id, singular: true
     relationship_by_property :rulers, property: :ruler_id
-    relationship_by_property :numismatic_monograms, property: :numismatic_monogram_ids
+    relationship_by_property :numismatic_monograms, property: :member_ids, model: Numismatics::Monogram
     relationship_by_property :file_sets, property: :member_ids, model: FileSet
     relationship_by_property :coins, property: :member_ids, model: Numismatics::Coin
     relationship_by_property :collections, property: :member_of_collection_ids