+++ app/resources/numismatics/issues/numismatics/issues_controller.rb
@@ -13,8 +13,6 @@ module Numismatics
     before_action :load_denominations, only: [:new, :edit]
     before_action :load_edges, only: [:new, :edit]
     before_action :load_metals, only: [:new, :edit]
-    before_action :load_monograms, only: [:new, :edit]
-    before_action :load_monogram_attributes, only: [:new, :edit]
     before_action :load_object_types, only: [:new, :edit]
     before_action :load_obverse_figures, only: [:new, :edit]
     before_action :load_obverse_orientations, only: [:new, :edit]
@@ -60,12 +58,6 @@ module Numismatics
 
     def load_metals
       @metals = @facet_values[:metal_ssim]
-    end
-
-    def load_monograms
-      @numismatic_monograms = query_service.find_all_of_model(model: Numismatics::Monogram).map(&:decorate)
-
-      return [] if @numismatic_monograms.to_a.blank?
     end
 
     def load_object_types
@@ -136,43 +128,5 @@ module Numismatics
         resource_class.new
       end
     end
-
-    private
-
-      def build_monogram_thumbnail_url(resource)
-        file_sets = resource.decorate.decorated_file_sets.reject { |fs| fs.thumbnail_id.nil? }
-
-        if file_sets.empty?
-          helpers.asset_url("default.png")
-        else
-          ManifestBuilder::ManifestHelper.new.manifest_image_thumbnail_path(file_sets.first)
-        end
-      end
-
-      # Generates the Vue JSON for the Numismatic Monogram membership component
-      # @return [Array<Hash>]
-      def load_monogram_attributes
-        numismatic_monogram_attributes = @numismatic_monograms.map do |monogram|
-          member_thumbnail_url = build_monogram_thumbnail_url(monogram)
-          member_url = solr_document_path(id: monogram.id)
-          member_monogram_ids = params[:id] ? resource.decorate.decorated_numismatic_monograms.map(&:id) : []
-
-          {
-            id: monogram.id.to_s,
-            url: member_url,
-            thumbnail: member_thumbnail_url,
-            title: monogram.decorate.first_title,
-            attached: member_monogram_ids.include?(monogram.id)
-          }
-        end
-
-        @numismatic_monogram_attributes = numismatic_monogram_attributes.sort do |u, v|
-          if u[:attached] <=> v[:attached]
-            0
-          else
-            v[:attached] && !u[:attached] ? 1 : -1
-          end
-        end
-      end
   end
 end