+++ spec/services/manifest_builder_spec.rb
@@ -917,12 +917,14 @@ RSpec.describe ManifestBuilder do
     subject(:manifest_builder) { described_class.new(query_service.find_by(id: numismatic_issue.id)) }
     let(:file) { fixture_file_upload("files/example.tif", "image/tiff") }
     let(:numismatic_issue) { FactoryBot.create_for_repository(:numismatic_issue) }
-    let(:change_set) { Numismatics::IssueChangeSet.new(numismatic_issue, member_ids: [coin1.id, coin2.id, coin3.id]) }
+    let(:change_set) { Numismatics::IssueChangeSet.new(numismatic_issue, member_ids: [coin1.id, coin2.id, coin3.id, monogram.id]) }
     let(:coin1) { FactoryBot.create_for_repository(:coin, files: [file1]) }
     let(:coin2) { FactoryBot.create_for_repository(:coin, files: [file2]) }
     let(:coin3) { FactoryBot.create_for_repository(:coin) }
+    let(:monogram) { FactoryBot.create_for_repository(:numismatic_monogram, files: [file3]) }
     let(:file1) { fixture_file_upload("files/abstract.tiff", "image/tiff") }
     let(:file2) { fixture_file_upload("files/abstract.tiff", "image/tiff") }
+    let(:file3) { fixture_file_upload("files/abstract.tiff", "image/tiff") }
     before do
       numismatic_issue
       change_set_persister.save(change_set: change_set)